patient registration with appointment and opdDetails

+++ src/main/java/com/hims/entity/repository/VisitRepository.java
@@ -3,5 +3,5 @@ package com.hims.entity.repository;
 import com.hims.entity.Visit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VisitRepository extends JpaRepository<Visit, Integer> {
+public interface VisitRepository extends JpaRepository<Visit, Long> {
 }
